backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==0.28.1
pandas==2.2.0
pyodbc==5.0.1
python-multipart==0.0.9
plotly==5.22.0
sqlalchemy==2.0.27
pydantic==2.6.1
tenacity==8.2.3
